refactor - move shared defs from subset to loader
