+++ scripts/schema/exclude_filter.py
@@ -1,7 +1,7 @@
 import glob
 import yaml
 import os
-from schema import subset_filter
+from schema import loader
 
 # This script should be run downstream of the subset filters - it takes
 # all ECS and custom fields already loaded by the latter and explicitly
@@ -18,44 +18,44 @@ def exclude(fields, exclude_file_globs, out_dir):
 
 
 def pop_field(fields, path):
-    '''pops a field from yaml derived dict using path derived from ordered list of nodes'''
+    """pops a field from yaml derived dict using path derived from ordered list of nodes"""
     node_path = path.copy()
     if node_path[0] in fields:
         if len(node_path) == 1:
             b4 = fields.copy()
-            print("Removed field {0}".format(str(fields.pop(node_path[0]).get("field_details").get("flat_name"))))
+            print('Removed field {0}'.format(str(fields.pop(node_path[0]).get('field_details').get('flat_name'))))
         else:
             inner_field = node_path.pop(0)
-            pop_field(fields[inner_field]["fields"], node_path)
+            pop_field(fields[inner_field]['fields'], node_path)
     else:
-        print("No match for exclusion:", ".".join([e for e in path]))
+        print('No match for exclusion:', '.'.join([e for e in path]))
 
 
 def exclude_trace_path(fields, item, path):
-    '''traverses paths to one or more nodes in a yaml derived dict'''
+    """traverses paths to one or more nodes in a yaml derived dict"""
     for list_item in item:
         node_path = path.copy()
-        node_path.append(list_item["name"])
-        if not "fields" in list_item:
+        node_path.append(list_item['name'])
+        if not 'fields' in list_item:
             pop_field(fields, node_path)
         else:
-            exclude_trace_path(fields, list_item["fields"], node_path)
+            exclude_trace_path(fields, list_item['fields'], node_path)
 
 
 def exclude_fields(fields, excludes):
-    '''Traverses fields and eliminates any field which matches the excludes'''
+    """Traverses fields and eliminates any field which matches the excludes"""
     if excludes:
         for ex_list in excludes:
             for item in ex_list:
-                #print("Removing: ", end='')
-                exclude_trace_path(fields, item["fields"], [item["name"]])
+                #print('Removing: ', end='')
+                exclude_trace_path(fields, item['fields'], [item['name']])
     return fields
 
 
 def load_exclude_definitions(file_globs):
     if not file_globs:
         return []
-    excludes = subset_filter.load_definitions(file_globs)
+    excludes = loader.load_definitions(file_globs)
     if not excludes:
         raise ValueError('--exclude specified, but no exclusions found in {}'.format(file_globs))
     return excludes
