+++ scripts/schema/loader.py
@@ -251,3 +251,30 @@ def merge_fields(a, b):
             a[key].setdefault('fields', {})
             a[key]['fields'] = merge_fields(a[key]['fields'], b[key]['fields'])
     return a
+
+
+def load_yaml_file(file_name):
+    with open(file_name) as f:
+        return yaml.safe_load(f.read())
+
+
+def eval_globs(globs):
+    '''Accepts an array of glob patterns or file names, returns the array of actual files'''
+    all_files = []
+    for g in globs:
+        new_files = glob.glob(g)
+        if len(new_files) == 0:
+            warn("{} did not match any files".format(g))
+        else:
+            all_files.extend(new_files)
+    return all_files
+
+
+def load_definitions(file_globs):
+    sets = []
+    for f in eval_globs(file_globs):
+        raw = load_yaml_file(f)
+        sets.append(raw)
+    return sets
+
+
